Read Temp Values

+++ Examples/MAX32665/BLE_tempbeacon/dats_api.h
@@ -65,7 +65,7 @@ void DatsHandlerInit(wsfHandlerId_t handlerId);
 /*************************************************************************************************/
 void DatsHandler(wsfEventMask_t event, wsfMsgHdr_t *pMsg);
 
-void updateTempValue(float tmpVal);
+void updateTempValue(uint16_t tmpVal);
 
 void WsfTimerInit_Sensor();
